Refactor / revamp
The whole codebase has been refactored to use new object types.
This is a breaking change as the references returned are now JNode interface instead of *JNode struct.
The new method is better standard Golang methodology.

In process of changes also added a number of unit tests and updated the parse to exit on error and return an error instead of panicking. The old "panic on error" functionality is still accessible on the Parse function. If panic is not desired, use ParseAll instead.