Fix perl array handling
An array from the root was not working due to an expectation to always start with {
Also, if there was whitespace before the { there would be an issue also.
Both issues have been corrected.

Besides this there was also an issue with objects within an array and that has now been fixed as well.