Many changes
C - Updated names of many functions and types to start with uj_ to reduce the pollution of C function namespace.

Updated the name overall internally of things to ujson instead of ujsonin to reflect that the library outputs json now instead of just parsing it. This isn't true of all the languages yet, but that proccess has begun.

Perl - Updated name to Ujson as well instead of Ujsonin.

Readme - Updated to reflect the current status of the project better.